scripts: Add a configuration and scripts for building with docker
Having a docker container that contains the cross compiler and host
dependencies allows for an easier setup than trying to ensure all the
right packages are installed.

This container builds top of tree alsa-lib and alsa-utils as distros
seem to be slow to update them with the needed topology features.

Signed-off-by: Dylan Reid <dgreid@chromium.org>